requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiosqlite==0.19.0
httpx==0.26.0
beautifulsoup4==4.12.3
anthropic==0.18.1
pydantic-settings==2.1.0
firecrawl-py==2.0.0
playwright==1.41.0
